refactor: create backend folder
