+++ src/app_model/backends/__init__.py
@@ -1,0 +1,3 @@
+"""Adapters for using the app_model with various backends."""
+
+# TODO: make a `use_app()` like adapter to easily switch?
