+++ src/app_model/backends/qt/_qaction.py
