+++ src/app_model/backends/qt/_util.py
