--- src/app_model/qt/__init__.py
@@ -1,4 +1,0 @@
-"""Qt objects for app_model."""
-from ._qmenu import QModelMenu
-
-__all__ = ["QModelMenu"]
